Pulled main function to root and reorganized imports
