+++ src/slpn_opt_uemsc_discovery.py
@@ -2,11 +2,12 @@ import pm4py
 import scipy
 
 from pm4py.objects.petri_net.utils import final_marking, initial_marking
-from src.slpn_visualiser import visualize_slpn, view
-from symbolic_conversion import calculate_inverse_poland_expression, get_inverse_poland_expression
 from pm4py.objects.log.importer.xes import importer as xes_importer
-from util import setup, get_slpn
-from slpn_exporter import export_slpn, export_slpn_xml
+
+from src.slpn_visualiser import visualize_slpn, view
+from src.symbolic_conversion import calculate_inverse_poland_expression, get_inverse_poland_expression
+from src.util import setup, get_slpn
+from src.slpn_exporter import export_slpn, export_slpn_xml
 
 
 def optimize_with_uemsc(log, pn, im, fm):
